+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 7.0.0
+
+#### Breaking
+- Updates the Braze Swift SDK bindings to require releases from the `13.0.0+` SemVer denomination.
+  - This allows compatibility with any version of the Braze SDK from `13.0.0` up to, but not including, `14.0.0`.
+  - Refer to the changelog entry for [`13.0.0`](https://github.com/braze-inc/braze-swift-sdk/blob/main/CHANGELOG.md#1300) for more information on potential breaking changes.
+
 ## 6.0.0
 
 #### Breaking